Replace consumption measure in charts with new aggConsumption.  Note
that the Hugo's original contains an additional 'pastConsumption' term,
which, if removed seems to kill the demand for credit.  I have left the
use of consumption in the model as Hugo intended it, as I do not
understand why he has this additional term yet.  But for accounting
purposes on the charts, the new aggConsumption term is used to show
actual consumption to gdp, etc.